backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai>=1.12.0
python-dotenv==1.0.0
moviepy==1.0.3
pydub==0.25.1
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx>=0.25.0
ffmpeg-python==0.2.0